+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Allowed numbers without integer part, i.e. `.123` or `.5e-4`
 - Allowed numbers and literals as property name in object literals, i.e. `{ 1: 'ok', null: 'ok' }`
 - Fixed suggestion in empty function body for new syntax, i.e. `group(=>)` will suggest between `=>` and `)`
+- Added `reduce()` method
 
 ## 1.0.0-alpha.13 (January 6, 2020)